Fix bad cast bug in Geohash::Reverse found by Christian Csar.  This was introduced in commit 4e95cacc4d73 (version 1.37) and affected all platforms where unsigned long is only 32-bits wide.

+++ src/Geohash.cpp
@@ -89,8 +89,8 @@ namespace GeographicLib {
     int s = 5 * (maxlen_ - len);
     ulon <<=     (s / 2);
     ulat <<= s - (s / 2);
-    lon = (unsigned long)(ulon) * loneps() - 180;
-    lat = (unsigned long)(ulat) * lateps() - 90;
+    lon = ulon * loneps() - 180;
+    lat = ulat * lateps() - 90;
   }
 
 } // namespace GeographicLib
